feat: add Topic model and endpoints
